confluence-page-creator/requirements.txt
playwright>=1.40.0
pyyaml>=6.0
asyncio
requests>=2.28.0